REEB-871
git-svn-id: http://svn.uhuila.cn/reeb/trunk@2626 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ module/consumer/app/models/consumer/UserInfo.java
@@ -63,7 +63,7 @@ public class UserInfo extends Model {
     public Date bindMobileAt;
 
     @Column(name = "total_points")
-    public Integer totalPoints;
+    public Long totalPoints;
 
     public UserInfo(User user) {
         this.user = user;
